generic definitons updated, new plot method included

+++ R/allClasses.R
@@ -21,5 +21,5 @@ setClass("Les",
                  limit="numeric",
                  nChr="integer",
                  minLength="integer",
-                 maxGap="integer")
+                 maxGap="numeric")
 )
